Separate home content in component
